GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - Started adding stuff for the metaboards and fixed the styles.
 (refs/gitbutler/Started-adding-stuff-for-the-metaboards-and-fixed-the-styles.-)
   branch head: edee436249be30dec03847782174df856d6b3064
 - main (refs/gitbutler/main)
   branch head: 214a9498acc39327b8394ffca0c6be81f0e30cf3
   - frontend/src/SvelteKanban.svelte
   - frontend/src/components/ListContainer.svelte
   - frontend/src/components/Item.svelte
   - frontend/src/components/List.svelte
   - frontend/src/components/Board.svelte
   - frontend/src/stores/metaboard.js
   - frontend/wailsjs/go/main/App.js
   - frontend/wailsjs/go/models.ts
   - frontend/wailsjs/go/main/App.d.ts
   - app.go
   - frontend/src/components/CommandBar.svelte
   - frontend/src/components/MetaBoardList.svelte
   - frontend/src/stores/styles.js

Your previous branch was: refs/heads/main

The sha for that commit was: 10ae2068199311600f16e2909883af309ae60d60

For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch